Refactored and Formatted files
