+++ frontend/package.json
@@ -1,42 +1,42 @@
 {
-    "name": "frontend",
-    "private": true,
-    "version": "0.1.0",
-    "type": "module",
-    "scripts": {
-        "dev": "vite",
-        "build": "tsc && vite build",
-        "preview": "vite preview",
-        "test": "vitest"
-    },
-    "dependencies": {
-        "bootstrap": "^5.2.3",
-        "react": "^18.2.0",
-        "react-bootstrap": "^2.7.2",
-        "react-bootstrap-icons": "^1.10.3",
-        "react-dom": "^18.2.0",
-        "react-hook-form": "^7.43.7",
-        "react-lines-ellipsis": "^0.15.3",
-        "react-router-dom": "^6.9.0"
-    },
-    "devDependencies": {
-        "@testing-library/jest-dom": "^5.16.5",
-        "@testing-library/react": "^14.0.0",
-        "@testing-library/user-event": "^14.4.3",
-        "@types/node": "^18.15.5",
-        "@types/react": "^18.0.28",
-        "@types/react-dom": "^18.0.11",
-        "@types/react-lines-ellipsis": "^0.15.1",
-        "@typescript-eslint/eslint-plugin": "^5.57.0",
-        "@typescript-eslint/parser": "^5.57.0",
-        "@vitejs/plugin-react-swc": "^3.2.0",
-        "eslint": "^8.37.0",
-        "eslint-plugin-react": "^7.32.2",
-        "jsdom": "^21.1.1",
-        "msw": "^1.1.0",
-        "prettier": "^2.8.7",
-        "typescript": "^5.0.2",
-        "vite": "^4.2.1",
-        "vitest": "^0.29.7"
-    }
+  "name": "frontend",
+  "private": true,
+  "version": "0.1.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview",
+    "test": "vitest"
+  },
+  "dependencies": {
+    "bootstrap": "^5.2.3",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.7.2",
+    "react-bootstrap-icons": "^1.10.3",
+    "react-dom": "^18.2.0",
+    "react-hook-form": "^7.43.7",
+    "react-lines-ellipsis": "^0.15.3",
+    "react-router-dom": "^6.9.0"
+  },
+  "devDependencies": {
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
+    "@types/node": "^18.15.5",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
+    "@types/react-lines-ellipsis": "^0.15.1",
+    "@typescript-eslint/eslint-plugin": "^5.57.0",
+    "@typescript-eslint/parser": "^5.57.0",
+    "@vitejs/plugin-react-swc": "^3.2.0",
+    "eslint": "^8.37.0",
+    "eslint-plugin-react": "^7.32.2",
+    "jsdom": "^21.1.1",
+    "msw": "^1.1.0",
+    "prettier": "^2.8.7",
+    "typescript": "^5.0.2",
+    "vite": "^4.2.1",
+    "vitest": "^0.29.7"
+  }
 }
