--- frontend/src/components/navbar/SignedInNavBar.tsx
@@ -1,45 +1,0 @@
-import { Button, Nav } from "react-bootstrap";
-import Navbar from "react-bootstrap/Navbar";
-import * as UserApi from "../../api/user";
-import { User } from "../../models/user";
-
-interface SignedInNavBarInProps {
-    user: User;
-    onSignOutSuccessful: () => void;
-}
-
-const SignedInNavBar = ({
-    user,
-    onSignOutSuccessful,
-}: SignedInNavBarInProps) => {
-    async function signout() {
-        try {
-            await UserApi.signout(user);
-            onSignOutSuccessful();
-        } catch (error) {
-            alert(error);
-            console.error(error);
-        }
-    }
-
-    return (
-        <>
-            <Nav className="ms-auto">
-                <Navbar.Text className="pe-2">
-                    Signed in as: <u>{user.username}</u>
-                </Navbar.Text>
-                <Nav.Item className="ms-1">
-                    <Button
-                        data-testid="Sign-Out"
-                        variant="danger"
-                        onClick={signout}
-                    >
-                        Sign Out
-                    </Button>
-                </Nav.Item>
-            </Nav>
-        </>
-    );
-};
-
-export default SignedInNavBar;
