+++ frontend/src/components/utils/UserNameWarning.tsx
@@ -1,9 +1,9 @@
 const UserNameWarning = () => {
-    return (
-        <ul>
-            <li>Username must be at least 8 characters long</li>
-        </ul>
-    );
+  return (
+    <ul>
+      <li>Username must be at least 8 characters long</li>
+    </ul>
+  );
 };
 
 export default UserNameWarning;
