+++ frontend/src/mocks/articles.ts
@@ -1,50 +1,52 @@
 export const articlesData = [
-    {
-        source: {
-            id: null,
-            name: "BBC News",
-        },
-        author: "https://www.facebook.com/bbcnews",
-        title: "North Korea fires long range missile ahead of Japan-South Korea talks - BBC",
-        description:
-            "The ICBM marks Pyongyang's fourth round of launches in a week as South Korea and Japan leaders meet.",
-        url: "https://www.bbc.com/news/world-asia-64972944",
-        urlToImage:
-            "https://ichef.bbci.co.uk/news/1024/branded_news/12C00/production/_129000867_295d9c5e11282629f1f6435246bd86930a0a9c040_99_5122_28815122x2881.jpg",
-        publishedAt: "2023-03-16T01:36:45Z",
-        content:
-            "North Korea has fired an intercontinental ballistic missile (ICBM) just hours before the leaders of South Korea and Japan are due to meet for a landmark summit.\r\nBoth Japanese and South Korean offici… [+1672 chars]",
+  {
+    source: {
+      id: null,
+      name: "BBC News",
     },
-    {
-        source: {
-            id: null,
-            name: "The Guardian",
-        },
-        author: "Guardian staff reporter",
-        title: "Nasa reveals new spacesuit for Artemis moon landing - The Guardian",
-        description:
-            "Outfits for planned lunar expedition more streamlined and flexible than Apollo suits, and a better fit for women",
-        url: "https://www.theguardian.com/science/2023/mar/16/nasa-reveals-new-spacesuit-for-artemis-moon-landing",
-        urlToImage:
-            "https://i.guim.co.uk/img/media/9616a27a67ad68df2482d4e3bbadec5aff0a50ba/1082_1380_6523_3914/master/6523.jpg?width=1200&height=630&quality=85&auto=format&fit=crop&overlay-align=bottom%2Cleft&overlay-width=100p&overlay-base64=L2ltZy9zdGF0aWMvb3ZlcmxheXMvdGctZGVmYXVsdC5wbmc&enable=upscale&s=320380f6ae10d00b6fc09d8a4292a971",
-        publishedAt: "2023-03-16T01:34:00Z",
-        content:
-            "The big, puffy white moonsuits worn by Neil Armstrong and his fellow Apollo astronauts a half-century ago are out of fashion. Lunar haute couture now calls for something more form-fitting and comfort… [+2810 chars]",
+    author: "https://www.facebook.com/bbcnews",
+    title:
+      "North Korea fires long range missile ahead of Japan-South Korea talks - BBC",
+    description:
+      "The ICBM marks Pyongyang's fourth round of launches in a week as South Korea and Japan leaders meet.",
+    url: "https://www.bbc.com/news/world-asia-64972944",
+    urlToImage:
+      "https://ichef.bbci.co.uk/news/1024/branded_news/12C00/production/_129000867_295d9c5e11282629f1f6435246bd86930a0a9c040_99_5122_28815122x2881.jpg",
+    publishedAt: "2023-03-16T01:36:45Z",
+    content:
+      "North Korea has fired an intercontinental ballistic missile (ICBM) just hours before the leaders of South Korea and Japan are due to meet for a landmark summit.\r\nBoth Japanese and South Korean offici… [+1672 chars]",
+  },
+  {
+    source: {
+      id: null,
+      name: "The Guardian",
     },
-    {
-        source: {
-            id: "cbs-news",
-            name: "CBS News",
-        },
-        author: "C Mandler",
-        title: 'Whoopi Goldberg apologizes for using a slur on "The View": "I should\'ve thought about it a little longer" - CBS News',
-        description:
-            "This is not the first time the actor and talk show host has come under fire for her comments on the show.",
-        url: "https://www.cbsnews.com/news/whoopi-goldberg-apologizes-ethnic-slur-the-view/",
-        urlToImage:
-            "https://assets3.cbsnewsstatic.com/hub/i/r/2023/03/15/84c25c93-775a-4347-8f35-ce24a7ac9fa7/thumbnail/1200x630g2/3c512322952a35f1a73cb25773c192e9/gettyimages-520717224.jpg",
-        publishedAt: "2023-03-16T00:47:33Z",
-        content:
-            "Whoopi Goldberg has apologized for using a word considered an ethnic slur on Wednesday's episode of ABC's \"The View.\" \r\nDuring a conversation on the broadcast discussing former President Donald Trump… [+2781 chars]",
+    author: "Guardian staff reporter",
+    title: "Nasa reveals new spacesuit for Artemis moon landing - The Guardian",
+    description:
+      "Outfits for planned lunar expedition more streamlined and flexible than Apollo suits, and a better fit for women",
+    url: "https://www.theguardian.com/science/2023/mar/16/nasa-reveals-new-spacesuit-for-artemis-moon-landing",
+    urlToImage:
+      "https://i.guim.co.uk/img/media/9616a27a67ad68df2482d4e3bbadec5aff0a50ba/1082_1380_6523_3914/master/6523.jpg?width=1200&height=630&quality=85&auto=format&fit=crop&overlay-align=bottom%2Cleft&overlay-width=100p&overlay-base64=L2ltZy9zdGF0aWMvb3ZlcmxheXMvdGctZGVmYXVsdC5wbmc&enable=upscale&s=320380f6ae10d00b6fc09d8a4292a971",
+    publishedAt: "2023-03-16T01:34:00Z",
+    content:
+      "The big, puffy white moonsuits worn by Neil Armstrong and his fellow Apollo astronauts a half-century ago are out of fashion. Lunar haute couture now calls for something more form-fitting and comfort… [+2810 chars]",
+  },
+  {
+    source: {
+      id: "cbs-news",
+      name: "CBS News",
     },
+    author: "C Mandler",
+    title:
+      'Whoopi Goldberg apologizes for using a slur on "The View": "I should\'ve thought about it a little longer" - CBS News',
+    description:
+      "This is not the first time the actor and talk show host has come under fire for her comments on the show.",
+    url: "https://www.cbsnews.com/news/whoopi-goldberg-apologizes-ethnic-slur-the-view/",
+    urlToImage:
+      "https://assets3.cbsnewsstatic.com/hub/i/r/2023/03/15/84c25c93-775a-4347-8f35-ce24a7ac9fa7/thumbnail/1200x630g2/3c512322952a35f1a73cb25773c192e9/gettyimages-520717224.jpg",
+    publishedAt: "2023-03-16T00:47:33Z",
+    content:
+      "Whoopi Goldberg has apologized for using a word considered an ethnic slur on Wednesday's episode of ABC's \"The View.\" \r\nDuring a conversation on the broadcast discussing former President Donald Trump… [+2781 chars]",
+  },
 ];
