+++ frontend/src/models/news.ts
@@ -1,21 +1,21 @@
 export interface News {
-    status: string;
-    totalResults: number;
-    articles: Article[];
+  status: string;
+  totalResults: number;
+  articles: Article[];
 }
 
 export interface Article {
-    source: Source;
-    author: null | string;
-    title: string;
-    description: string;
-    url: string;
-    urlToImage: string;
-    publishedAt: Date;
-    content: string;
+  source: Source;
+  author: null | string;
+  title: string;
+  description: string;
+  url: string;
+  urlToImage: string;
+  publishedAt: Date;
+  content: string;
 }
 
 export interface Source {
-    id: null | string;
-    name: string;
+  id: null | string;
+  name: string;
 }
