+++ frontend/src/pages/HomePage.tsx
@@ -1,25 +1,21 @@
 import { Container } from "react-bootstrap";
-import SignedInView from "../components/views/SignedInView";
-import SignedOutView from "../components/views/SignedOutView";
+import HomeView from "../components/view/HomeView";
+import UserView from "../components/view/UserView";
 import { User } from "../models/user";
 import styles from "../styles/NewsPage.module.css";
 
 interface HomePageProps {
-    signedInUser: User | null;
+  signedInUser: User | null;
 }
 
 const HomePage = ({ signedInUser }: HomePageProps) => {
-    return (
-        <>
-            <Container className={styles.pageContainer}>
-                {signedInUser ? (
-                    <SignedInView user={signedInUser} />
-                ) : (
-                    <SignedOutView />
-                )}
-            </Container>
-        </>
-    );
+  return (
+    <>
+      <Container className={styles.pageContainer}>
+        {signedInUser ? <UserView user={signedInUser} /> : <HomeView />}
+      </Container>
+    </>
+  );
 };
 
 export default HomePage;
