+++ frontend/src/styles/Article.module.css
@@ -1,4 +1,4 @@
 .image {
-    height: 350px;
-    object-fit: cover;
+  height: 350px;
+  object-fit: cover;
 }
