+++ frontend/src/styles/NewsPage.module.css
@@ -1,3 +1,3 @@
 .pageContainer {
-    margin: 2rem auto;
+  margin: 2rem auto;
 }
