+++ frontend/src/utils/HttpErrors.ts
@@ -1,8 +1,8 @@
 class HttpErrors extends Error {
-    constructor(message?: string) {
-        super(message);
-        this.name = this.constructor.name;
-    }
+  constructor(message?: string) {
+    super(message);
+    this.name = this.constructor.name;
+  }
 }
 
 /**
