+++ frontend/vite.config.ts
@@ -3,39 +3,35 @@ import { defineConfig } from "vitest/config";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-    plugins: [react()],
-    test: {
-        globals: true,
-        setupFiles: "src/setupTests.ts",
-        environment: "jsdom",
-    },
-    server: {
-        proxy: {
-            "/api": {
-                target: "http://localhost:8080",
-                changeOrigin: true,
-                secure: false,
-                ws: true,
-                configure: (proxy, _options) => {
-                    proxy.on("error", (err, _req, _res) => {
-                        console.log("proxy error", err);
-                    });
-                    proxy.on("proxyReq", (proxyReq, req, _res) => {
-                        console.log(
-                            "Sending Request to the Target:",
-                            req.method,
-                            req.url
-                        );
-                    });
-                    proxy.on("proxyRes", (proxyRes, req, _res) => {
-                        console.log(
-                            "Received Response from the Target:",
-                            proxyRes.statusCode,
-                            req.url
-                        );
-                    });
-                },
-            },
+  plugins: [react()],
+  test: {
+    globals: true,
+    setupFiles: "src/setupTests.ts",
+    environment: "jsdom",
+  },
+  server: {
+    proxy: {
+      "/api": {
+        target: "http://localhost:8080",
+        changeOrigin: true,
+        secure: false,
+        ws: true,
+        configure: (proxy, _options) => {
+          proxy.on("error", (err, _req, _res) => {
+            console.log("proxy error", err);
+          });
+          proxy.on("proxyReq", (proxyReq, req, _res) => {
+            console.log("Sending Request to the Target:", req.method, req.url);
+          });
+          proxy.on("proxyRes", (proxyRes, req, _res) => {
+            console.log(
+              "Received Response from the Target:",
+              proxyRes.statusCode,
+              req.url,
+            );
+          });
         },
+      },
     },
+  },
 });
